requirements.txt
Django==3.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
pytz==2019.3
Faker==3.0.0